fix(site): char-level animation, drain at StreamingOutput level
Three changes:

1. Back to sep:char (duration 60ms, stagger 12ms). Word-level
   felt chunky. A 10-char SSE burst takes ~120ms+60ms to reveal,
   which fills the typical 100-200ms gap between deliveries.

2. Move the animation drain from Response up to StreamingOutput.
   The chop was caused by StreamingOutput unmounting entirely when
   liveStatus.phase goes idle — the message re-renders as a static
   ChatMessageItem, destroying all in-flight CSS animations. The
   new useStreamDrain hook detects the streaming→non-streaming
   transition and keeps StreamingOutput mounted, listening for
   animationend events on the container. When no more fire within
   60ms, the drain ends and the component unmounts cleanly.

3. Simplify Response back to a pure passthrough — no drain logic,
   no merged refs. The isAnimating prop goes straight to Streamdown.
